MSI: Indexation improvement
-- refactoring
