+++ app/code/Magento/Inventory/Indexer/IndexStructure.php
@@ -3,15 +3,13 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
+declare(strict_types=1);
 
-namespace Magento\Inventory\Indexer\StockItem;
+namespace Magento\Inventory\Indexer;
 
 use Magento\Framework\App\ResourceConnection;
 use Magento\Framework\DB\Ddl\Table;
 use Magento\Framework\Exception\StateException;
-use Magento\Inventory\Indexer\IndexName;
-use Magento\Inventory\Indexer\IndexNameResolverInterface;
-use Magento\Inventory\Indexer\IndexStructureInterface;
 
 /**
  * @inheritdoc
