--- app/code/Magento/Inventory/Indexer/Source.php
@@ -1,100 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-namespace Magento\Inventory\Indexer;
-
-use Magento\Inventory\Indexer\Source\GetPartialReindexData;
-use Magento\Inventory\Indexer\StockItem\IndexDataProvider;
-use Magento\InventoryApi\Api\Data\StockInterface;
-
-/**
- * @inheritdoc
- */
-class Source implements SourceIndexerInterface
-{
-    /**
-     * @var GetPartialReindexData
-     */
-    private $getPartialReindexData;
-
-    /**
-     * @var IndexStructureInterface
-     */
-    private $indexStructure;
-
-    /**
-     * @var IndexHandlerInterface
-     */
-    private $indexHandler;
-
-    /**
-     * @var IndexDataProvider
-     */
-    private $indexDataProvider;
-
-    /**
-     * @var IndexNameBuilder
-     */
-    private $indexNameBuilder;
-
-    /**
-     * @var StockItemIndexerInterface
-     */
-    private $stockItemIndexer;
-
-    /**
-     * @param GetPartialReindexData $getPartialReindexData
-     * @param IndexStructureInterface $indexStructureHandler
-     * @param IndexHandlerInterface $indexHandler
-     * @param IndexDataProvider $indexDataProvider
-     * @param IndexNameBuilder $indexNameBuilder
-     * @param StockItemIndexerInterface $stockItemIndexer
-     */
-    public function __construct(
-        GetPartialReindexData $getPartialReindexData,
-        IndexStructureInterface $indexStructureHandler,
-        IndexHandlerInterface $indexHandler,
-        IndexDataProvider $indexDataProvider,
-        IndexNameBuilder $indexNameBuilder,
-        StockItemIndexerInterface $stockItemIndexer
-    ) {
-        $this->getPartialReindexData = $getPartialReindexData;
-        $this->indexStructure = $indexStructureHandler;
-        $this->indexHandler = $indexHandler;
-        $this->indexDataProvider = $indexDataProvider;
-        $this->indexNameBuilder = $indexNameBuilder;
-        $this->stockItemIndexer = $stockItemIndexer;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function executeFull()
-    {
-        $this->stockItemIndexer->executeFull();
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function executeRow($sourceId)
-    {
-        $this->executeList([$sourceId]);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function executeList(array $sourceIds)
-    {
-        $stockIds = [];
-        $stockList = $this->getPartialReindexData->execute($sourceIds);
-        foreach ($stockList as $stockData) {
-            $stockIds[] = $stockData[StockInterface::STOCK_ID];
-        }
-        $this->stockItemIndexer->executeList($stockIds);
-    }
-}
