+++ app/code/Magento/Inventory/Indexer/Source/SourceIndexerInterface.php
@@ -3,19 +3,22 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\Inventory\Indexer;
+declare(strict_types=1);
+
+namespace Magento\Inventory\Indexer\Source;
 
 use Magento\Framework\Indexer\ActionInterface;
 
 /**
+ * Source indexer
  * Extension point for indexation
  *
  * @api
  */
-interface StockItemIndexerInterface extends ActionInterface
+interface SourceIndexerInterface extends ActionInterface
 {
     /**
      * Indexer ID in configuration
      */
-    const INDEXER_ID = 'inventory_stock_item';
+    const INDEXER_ID = 'inventory_source';
 }
