+++ app/code/Magento/Inventory/Indexer/SourceItem/SkuListInStockToUpdate.php
@@ -3,7 +3,9 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\Inventory\Indexer\StockItem;
+declare(strict_types=1);
+
+namespace Magento\Inventory\Indexer\SourceItem;
 
 /**
  * Represents relation between stock and sku list
