+++ app/code/Magento/Inventory/Indexer/SourceItem/SourceItemIndexerInterface.php
@@ -3,19 +3,22 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\Inventory\Indexer;
+declare(strict_types=1);
+
+namespace Magento\Inventory\Indexer\SourceItem;
 
 use Magento\Framework\Indexer\ActionInterface;
 
 /**
+ * Source Item indexer
  * Extension point for indexation
  *
  * @api
  */
-interface SourceIndexerInterface extends ActionInterface
+interface SourceItemIndexerInterface extends ActionInterface
 {
     /**
      * Indexer ID in configuration
      */
-    const INDEXER_ID = 'inventory_source';
+    const INDEXER_ID = 'inventory_source_item';
 }
