+++ app/code/Magento/Inventory/etc/indexer.xml
@@ -6,19 +6,19 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Indexer/etc/indexer.xsd">
-    <indexer id="inventory_stock_item" view_id="inventory_stock_item" class="Magento\Inventory\Indexer\StockItem">
-        <title translate="true">Stock Item</title>
-        <description translate="true">Rebuild Stock Item index (MSI)</description>
+    <indexer id="inventory_stock" view_id="inventory_stock" class="Magento\Inventory\Indexer\Stock\StockIndexer">
+        <title translate="true">Stocks</title>
+        <description translate="true">Rebuild Stock index (MSI)</description>
     </indexer>
-    <indexer id="inventory_source" view_id="inventory_source" class="Magento\Inventory\Indexer\Source" hidden="true">
-        <title translate="true">Source</title>
+    <indexer id="inventory_source" view_id="inventory_source" class="Magento\Inventory\Indexer\Source\SourceIndexer" hidden="true">
+        <title translate="true">Sources</title>
         <description translate="true">Rebuild Source index (MSI)</description>
         <dependencies>
             <indexer id="inventory_stock_item" />
         </dependencies>
     </indexer>
-    <indexer id="inventory_source_item" view_id="inventory_source_item" class="Magento\Inventory\Indexer\SourceItem" hidden="true">
-        <title translate="true">Source Item</title>
+    <indexer id="inventory_source_item" view_id="inventory_source_item" class="Magento\Inventory\Indexer\SourceItem\SourceItem" hidden="true">
+        <title translate="true">Source Items</title>
         <description translate="true">Rebuild Source Item index (MSI)</description>
         <dependencies>
             <indexer id="inventory_stock_item" />
