+++ app/code/Magento/Inventory/etc/mview.xml
@@ -6,7 +6,7 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Mview/etc/mview.xsd">
-    <view id="inventory_stock_item" class="Magento\Inventory\Indexer\StockItem\Mview\Action" group="indexer">
+    <view id="inventory_stock" class="Magento\Inventory\Indexer\Stock\Mview\Action" group="indexer">
         <subscriptions>
             <table name="inventory_source_stock_link" entity_column="stock_id" />
         </subscriptions>
